104-fibonacci.c

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -10,18 +10,18 @@ int main(void)
 int64_t n1 = 1, n2 = 2, n3;
 
 int i, n = 98;
-printf("%lld, %lld", n1, n2);
+printf("%lld, %lld", (long long int)n1, (long long int)n2);
 for (i = 3; i <= n; ++i)
 {
 
 n3 = n1 + n2;
 if (i < n)
 {
-printf(", %lld", n3);
+printf(", %lld", (long long int)n3);
 }
 else
 {
-printf(", %lld", n3);
+printf(", %lld", (long long int)n3);
 printf("\n");
 }
 n1 = n2;
